feat(background-image): Add background img to page
Add a background image (of the calanques in Marseille!) to index.html.
According to the style spec, the image needs to be on the left side of
the index page taking up about 30% of the width of the page; thus add
the required divs and CSS styling to accomplish that task. Note: the
background image isn't visible currently because of collapsing divs but
will be visible once the divs are filled with appropriate content.