feat(styles): Add background and font colors
Add a form-container class to the appropriate div in the index page for
styling purposes. In the styles file, add background color #00001a to
style the background a dark navy color that goes with the dark color of
the water in the calanques background image. Style the font a beige
color that goes with the rocky surfaces in the calanques photo and also
provides contrast with the background color for ease-of-reading.