feat(styles): Add CSS styling match spec
Add margins, padding, reduce font size of input field labels, add a logo
and a semi-transparent div behind it as the foreground of the background
image, organize the input fields into two columns. Use CSS variables for
light and dark colors for improved maintainability/ease of changing
design in the future.